refactor: unify navigation hooks for ReferFriends feature
- Replaced the existing navigation hooks for ReferAFriend with a unified hook, useNavigateToReferFriends, to streamline navigation logic.
- Updated BreadcrumbSection and useRedirectWhenNotLoggedIn to utilize the new hook, enhancing code maintainability.
- Removed obsolete hooks related to InviteReward navigation, consolidating functionality for better clarity and efficiency.
