+++ packages/kit/src/views/ReferFriends/components/BreadcrumbSection.tsx
@@ -4,7 +4,7 @@ import { Breadcrumb } from '@onekeyhq/components';
 import type { IBreadcrumbItem } from '@onekeyhq/components/src/content/Breadcrumb';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 
-import { useReplaceToReferAFriend } from '../pages/ReferAFriend/hooks/useNavigateToReferAFriend';
+import { useReplaceToReferFriends } from '../hooks/useNavigateToReferFriends';
 
 export interface IBreadcrumbSectionProps {
   secondItemLabel: string;
@@ -14,13 +14,13 @@ export function BreadcrumbSection({
   secondItemLabel,
 }: IBreadcrumbSectionProps) {
   const intl = useIntl();
-  const replaceToReferAFriend = useReplaceToReferAFriend();
+  const replaceToReferFriends = useReplaceToReferFriends();
 
   const breadcrumbItems: IBreadcrumbItem[] = [
     {
       label: intl.formatMessage({ id: ETranslations.global_overview }),
       onClick: () => {
-        replaceToReferAFriend({});
+        void replaceToReferFriends({});
       },
     },
     {
