+++ packages/kit/src/views/ReferFriends/hooks/useRedirectWhenNotLoggedIn.ts
@@ -2,15 +2,16 @@ import { useEffect } from 'react';
 
 import { usePrimePersistAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
 
-import { useReplaceToReferAFriend } from '../pages/ReferAFriend/hooks/useNavigateToReferAFriend';
+import { useReplaceToReferFriends } from './useNavigateToReferFriends';
 
 /**
- * Hook to redirect to ReferAFriend page when user is not logged in
+ * Hook to redirect to appropriate ReferFriends page when user is not logged in
  * - Monitors login status in real-time using Jotai atom
  * - Redirects using replace method to prevent navigation back
+ * - Uses unified navigation that checks login status
  */
 export function useRedirectWhenNotLoggedIn() {
-  const replaceToReferAFriend = useReplaceToReferAFriend();
+  const replaceToReferFriends = useReplaceToReferFriends();
   const [primeAtom] = usePrimePersistAtom();
 
   // Monitor login status changes in real-time
@@ -18,11 +19,11 @@ export function useRedirectWhenNotLoggedIn() {
     const isLoggedIn = primeAtom.isLoggedIn && primeAtom.isLoggedInOnServer;
 
     if (!isLoggedIn) {
-      replaceToReferAFriend();
+      void replaceToReferFriends();
     }
   }, [
     primeAtom.isLoggedIn,
     primeAtom.isLoggedInOnServer,
-    replaceToReferAFriend,
+    replaceToReferFriends,
   ]);
 }
