+++ packages/kit/src/views/ReferFriends/pages/InviteReward/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './useNavigateToInviteReward';
+// No exports - InviteReward page uses hooks from the shared ReferFriends hooks folder
