+++ packages/kit/src/views/ReferFriends/pages/ReferAFriend/hooks/index.ts
@@ -1,3 +1,2 @@
-export * from './useNavigateToReferAFriend';
-export * from './useReferAFriendData';
 export * from './useLoginStatusChange';
+export * from './useReferAFriendData';
